+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/BotBuilder.kt
@@ -4,8 +4,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.utils.TelegramAPIUrlsKeeper
 import io.ktor.client.HttpClient
 import io.ktor.client.HttpClientConfig
-import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.ProxyConfig
+import io.ktor.client.engine.*
 
 /**
  * @param proxy Standard ktor [ProxyConfig]
@@ -14,20 +13,30 @@ import io.ktor.client.engine.ProxyConfig
  */
 data class BotBuilder internal constructor(
     var proxy: ProxyConfig? = null,
+    @Deprecated("ktorClientEngineFactory parameter will be used preferable. In future this parameter will be removed")
     var ktorClientEngine: HttpClientEngine? = null,
+    var ktorClientEngineFactory: HttpClientEngineFactory<out HttpClientEngineConfig>? = null,
     var ktorClientConfig: (HttpClientConfig<*>.() -> Unit) ? = null
 ) {
-    internal fun createHttpClient(): HttpClient = ktorClientEngine ?.let { engine ->
+    internal fun createHttpClient(): HttpClient = ktorClientEngineFactory ?.let {
+        HttpClient(
+            it.create {
+                this@create.proxy = this@BotBuilder.proxy ?: this@create.proxy
+            }
+        ) {
+            ktorClientConfig ?.let { it() }
+        }
+    } ?: ktorClientEngine ?.let { engine ->
         HttpClient(engine) {
             ktorClientConfig ?.let { it() }
             engine {
-                proxy = this@BotBuilder.proxy ?: proxy
+                this@engine.proxy = this@BotBuilder.proxy ?: this@engine.proxy
             }
         }
     } ?: HttpClient {
         ktorClientConfig ?.let { it() }
         engine {
-            proxy = this@BotBuilder.proxy ?: proxy
+            this@engine.proxy = this@BotBuilder.proxy ?: this@engine.proxy
         }
     }
 }