+++ CHANGELOG.md
@@ -68,6 +68,8 @@
 * `TelegramBotAPI-extensions-api`:
     * A lot of `RequesstExecutor#getChat` extensions was added for more explicit types showing
     * New `RequesstExecutor#setMyCommands` extension was added
+    * New field `BotBuilder#ktorClientEngineFactory` introduced
+        * Field `BotBuilder#ktorClientEngine` now is deprecated
 * `TelegramBotAPI-extensions-utils`:
     * `safely` function was introduced. It is in `PreviewFeature` state currently
     * `makeLinkToMessage` extensions has been added